diagnostics: unify Oracle-aware verify runner across CLI and audit
aver audit went through a stripped-down verify runner in
diagnostics::verify_run that synthesized bare helper fns — no `given`
stub installation, no .trace / .result projection stripping, no
trace-event post-processing. On the same source aver verify passed
31/31 while aver audit reported 27 failures. Release blocker for
anyone trusting audit output.

Extract the full Oracle-aware runner from src/main/commands.rs into
src/diagnostics/vm_verify.rs as pub API:
- run_verify_for_items_vm (disk base_dir)
- run_verify_for_items_vm_with_loaded (virtual fs)

Both accept an optional pre-loaded ProjectConfig so CLI and audit
apply identical runtime policy. The runner owns plan synthesis with
trace-projection stripping, per-case oracle stub installation,
trace collection with root_fn_id scoping, and the apply_trace_projection
post-processor for .trace.{length,event,contains} / .trace.group(N)
shapes.

verify_run.rs now delegates to vm_verify and maps the runner's
VerifyResults onto canonical diagnostics. commands.rs call sites
call the same lib function with load_runtime_policy plugged in for
the config. 962 lines of duplicated verify logic deleted.
