+++ src/main/commands.rs
@@ -2,17 +2,12 @@ use std::collections::{HashMap, HashSet};
 use std::fs;
 use std::path::{Path, PathBuf};
 use std::process;
-use std::sync::Arc as Rc;
 use std::time::{SystemTime, UNIX_EPOCH};
 
 use colored::Colorize;
 
-use aver::ast::{
-    Expr, FnBody, FnDef, Pattern, Spanned, Stmt, TopLevel, TypeDef, VerifyBlock, VerifyKind,
-};
-use aver::checker::{
-    CheckFinding, VerifyResult, expr_to_str, index_decisions, merge_verify_blocks,
-};
+use aver::ast::{Expr, FnDef, Pattern, Spanned, Stmt, TopLevel, TypeDef, VerifyKind};
+use aver::checker::{CheckFinding, VerifyResult, index_decisions};
 use aver::codegen;
 use aver::codegen::ModuleInfo;
 use aver::codegen::lean as lean_codegen;
@@ -23,7 +18,6 @@ use aver::source::{find_module_file, require_module_declaration};
 use aver::tco;
 use aver::types::checker::run_type_check_full;
 use aver::types::{Type, parse_type_str};
-use aver::value::{Value, aver_repr};
 use aver::verify_law::{
     collect_contextual_helper_law_hints, collect_missing_helper_law_hints,
     contextual_helper_law_message, missing_helper_law_message,
@@ -1266,7 +1260,16 @@ pub(super) fn cmd_run_vm(
 
     if run_verify_blocks {
         println!();
-        match run_verify_for_items_vm(items, &module_root, file) {
+        let cfg = load_runtime_policy(&module_root).unwrap_or_else(|e| {
+            eprintln!("{}", e.red());
+            process::exit(1);
+        });
+        match aver::diagnostics::vm_verify::run_verify_for_items_vm(
+            items,
+            cfg,
+            Some(&module_root),
+            file,
+        ) {
             Ok(results) => {
                 let failed: usize = results.iter().map(|r| r.failed).sum();
                 let file_results = vec![VerifyFileResult {
@@ -2503,969 +2506,6 @@ pub(super) fn cmd_check(
     }
 }
 
-struct VmVerifyCaseFns {
-    left: String,
-    right: String,
-    guard: Option<String>,
-}
-
-struct VmVerifyPlan {
-    block: VerifyBlock,
-    cases: Vec<VmVerifyCaseFns>,
-}
-
-enum VmVerifyEval {
-    Value(Value),
-    ErrProp(Value),
-}
-
-fn make_verify_vm_helper(
-    name: String,
-    line: usize,
-    expr: Spanned<Expr>,
-    wrap_result: bool,
-) -> TopLevel {
-    let body_expr = if wrap_result {
-        Spanned::new(
-            Expr::Constructor("Result.Ok".to_string(), Some(Box::new(expr))),
-            line,
-        )
-    } else {
-        expr
-    };
-
-    TopLevel::FnDef(FnDef {
-        name,
-        line,
-        params: vec![],
-        return_type: "Unit".to_string(),
-        effects: vec![],
-        desc: None,
-        body: Rc::new(FnBody::from_expr(body_expr)),
-        resolution: None,
-    })
-}
-
-fn build_verify_vm_plans(
-    items: &mut Vec<TopLevel>,
-    verify_blocks: &[VerifyBlock],
-) -> Vec<VmVerifyPlan> {
-    let mut plans = Vec::with_capacity(verify_blocks.len());
-
-    for (block_idx, block) in verify_blocks.iter().enumerate() {
-        let mut case_plans = Vec::with_capacity(block.cases.len());
-        let sample_guards = match &block.kind {
-            VerifyKind::Law(law) => Some(&law.sample_guards),
-            VerifyKind::Cases => None,
-        };
-
-        for (case_idx, (left_expr, right_expr)) in block.cases.iter().cloned().enumerate() {
-            let prefix = format!("__verify_{}_{}_{}", block.fn_name, block_idx, case_idx);
-            let left_name = format!("{}_left", prefix);
-            let right_name = format!("{}_right", prefix);
-            // Oracle v1: in trace-aware blocks, LHS trace-projection
-            // shapes are rewritten to the underlying fn() call so the
-            // generated helper just runs the fn (which populates the
-            // trace-events buffer via VM hooks). The projection value
-            // (contains / event / length / result) is then computed from
-            // the collected events in `apply_trace_projection` after
-            // helper eval, replacing the helper's raw return.
-            let left_expr = if block.trace {
-                strip_trace_projection(left_expr)
-            } else {
-                left_expr
-            };
-            items.push(make_verify_vm_helper(
-                left_name.clone(),
-                block.line,
-                left_expr,
-                true,
-            ));
-            items.push(make_verify_vm_helper(
-                right_name.clone(),
-                block.line,
-                right_expr,
-                true,
-            ));
-
-            let guard_name = sample_guards
-                .and_then(|guards| guards.get(case_idx))
-                .cloned()
-                .map(|guard_expr| {
-                    let name = format!("{}_guard", prefix);
-                    items.push(make_verify_vm_helper(
-                        name.clone(),
-                        block.line,
-                        guard_expr,
-                        false,
-                    ));
-                    name
-                });
-
-            case_plans.push(VmVerifyCaseFns {
-                left: left_name,
-                right: right_name,
-                guard: guard_name,
-            });
-        }
-
-        plans.push(VmVerifyPlan {
-            block: block.clone(),
-            cases: case_plans,
-        });
-    }
-
-    plans
-}
-
-/// Oracle v1: strip a trace projection from a verify-trace LHS,
-/// returning the underlying `fn(args)` call that should be evaluated to
-/// populate the trace-events buffer. Handles:
-///
-/// - `fn(args).result`
-/// - `fn(args).trace.contains(_)`
-/// - `fn(args).trace.event(_)`
-/// - `fn(args).trace.length()`
-/// - `fn(args).trace`
-///
-/// Other expression shapes pass through unchanged.
-fn strip_trace_projection(expr: Spanned<Expr>) -> Spanned<Expr> {
-    if let Some(fc) = find_trace_underlying_fn_call(&expr.node) {
-        return fc;
-    }
-    expr
-}
-
-fn find_trace_underlying_fn_call(node: &Expr) -> Option<Spanned<Expr>> {
-    match node {
-        // fn(args).result
-        Expr::Attr(inner, field) if field == "result" => {
-            if matches!(&inner.node, Expr::FnCall(_, _)) {
-                Some((**inner).clone())
-            } else {
-                None
-            }
-        }
-        // fn(args).trace
-        Expr::Attr(inner, field) if field == "trace" => {
-            if matches!(&inner.node, Expr::FnCall(_, _)) {
-                Some((**inner).clone())
-            } else {
-                None
-            }
-        }
-        // fn(args).trace.<method>(...)            (flat)
-        // fn(args).trace.group(N).<method>(...)   (one tree-nav stage)
-        Expr::FnCall(callee, _) => {
-            let Expr::Attr(inner_attr, _) = &callee.node else {
-                return None;
-            };
-            // Allow optional `.group(N)` / `.branch(idx)` tree-nav
-            // stages between `.trace` and the terminal projection
-            // method. We walk inward until we hit `.trace`, then grab
-            // the wrapped fn-call.
-            let (scope, _, _) = peel_tree_nav_stages(inner_attr);
-            let Expr::Attr(fn_call_box, trace_field) = &scope.node else {
-                return None;
-            };
-            if trace_field == "trace" && matches!(&fn_call_box.node, Expr::FnCall(_, _)) {
-                return Some((**fn_call_box).clone());
-            }
-            None
-        }
-        _ => None,
-    }
-}
-
-/// Oracle v1: replace the helper's raw return with the projected value
-/// from the LHS's trace-API shape, if applicable.
-///
-/// - `.result` stays as-is (helper already evaluates the fn_call; its
-///   return IS the result).
-/// - `.trace.length()` → `Int` (events count).
-/// - `.trace.contains(event_lit)` → `Bool` (structural match).
-/// - `.trace.event(k)` → `Option<EffectEvent>`.
-/// - `.trace` → Trace record (wraps events list).
-///
-/// `original_lhs` is the un-stripped expression used to detect shape.
-/// `collected` is the list of EffectEvent records from the trace buffer.
-fn apply_trace_projection(
-    helper_result: Result<VmVerifyEval, String>,
-    original_lhs: &Spanned<Expr>,
-    collected: &[Value],
-    coords: &[aver::vm::runtime::TraceCoord],
-    trace_mode: bool,
-) -> Result<VmVerifyEval, String> {
-    if !trace_mode {
-        return helper_result;
-    }
-    let helper_result = helper_result?;
-    // Keep raw return for .result / default path.
-    if matches!(
-        &original_lhs.node,
-        Expr::Attr(_, f) if f == "result"
-    ) {
-        return Ok(helper_result);
-    }
-    match &original_lhs.node {
-        // fn(args).trace — return Trace record (whole buffer, flat)
-        Expr::Attr(inner, field)
-            if field == "trace" && matches!(&inner.node, Expr::FnCall(_, _)) =>
-        {
-            Ok(VmVerifyEval::Value(build_trace_record(collected)))
-        }
-        // fn(args).trace.<method>(...) and
-        // fn(args).trace.group(N).<method>(...)
-        Expr::FnCall(callee, args) => {
-            let Expr::Attr(inner_attr, method) = &callee.node else {
-                return Ok(helper_result);
-            };
-            // Peel the optional tree-nav stages. `inner_attr` may be:
-            //   fn_call.trace                       — flat
-            //   fn_call.trace.group(N)              — one stage
-            //   fn_call.trace.group(N).branch(idx)  — two stages
-            let (scope, group_id, branch_idx) = peel_tree_nav_stages(inner_attr);
-            let Expr::Attr(fn_call_box, trace_field) = &scope.node else {
-                return Ok(helper_result);
-            };
-            if trace_field != "trace" || !matches!(&fn_call_box.node, Expr::FnCall(_, _)) {
-                return Ok(helper_result);
-            }
-            // Filter events by group + branch when each stage is
-            // present. User-visible `N` is 0-based in source order;
-            // the replay scope assigns internal group ids starting at
-            // 1, so source index 0 → replay group_id 1. Branch idx
-            // maps directly (0-based in both source and runtime).
-            let filtered: Vec<&Value> = collected
-                .iter()
-                .zip(coords.iter())
-                .filter(|(_, c)| match group_id {
-                    Some(n) => c.group_id == Some(n + 1),
-                    None => true,
-                })
-                .filter(|(_, c)| match branch_idx {
-                    Some(b) => c.branch_idx == Some(b),
-                    None => true,
-                })
-                .map(|(ev, _)| ev)
-                .collect();
-            match method.as_str() {
-                "length" => Ok(VmVerifyEval::Value(Value::Int(filtered.len() as i64))),
-                "event" => {
-                    if args.len() != 1 {
-                        return Err(format!("Trace.event(k) expects 1 arg, got {}", args.len()));
-                    }
-                    let k = match &args[0].node {
-                        Expr::Literal(aver::ast::Literal::Int(i)) if *i >= 0 => *i as usize,
-                        _ => {
-                            return Err(
-                                "Trace.event(k) requires a literal non-negative Int for v0"
-                                    .to_string(),
-                            );
-                        }
-                    };
-                    let v = match filtered.get(k) {
-                        Some(ev) => Value::Some(Box::new((**ev).clone())),
-                        None => Value::None,
-                    };
-                    Ok(VmVerifyEval::Value(v))
-                }
-                "contains" => {
-                    if args.len() != 1 {
-                        return Err(format!(
-                            "Trace.contains(x) expects 1 arg, got {}",
-                            args.len()
-                        ));
-                    }
-                    // Oracle v1: `.contains(x)` accepts two shapes:
-                    //   - `Effect.method(args…)` — full event literal,
-                    //     matches method + args.
-                    //   - `Effect.method` (no parens) — method-only
-                    //     predicate, matches any event with that method
-                    //     regardless of args. Useful for "did this
-                    //     effect fire at all" checks.
-                    let arg = &args[0];
-                    let method_only_name = effect_method_ref_name(arg);
-                    let hit = if let Some(method_name) = method_only_name {
-                        filtered
-                            .iter()
-                            .any(|ev| effect_event_has_method(ev, &method_name))
-                    } else {
-                        let expected = match expr_to_effect_event(arg) {
-                            Some(ev) => ev,
-                            None => {
-                                return Err(
-                                    "Trace.contains(x): argument must be either an effect-method \
-                                     reference (e.g. `Console.print`) or an effect-method call \
-                                     (e.g. `Console.print(\"...\")`) that elaborates to an event literal"
-                                        .to_string(),
-                                );
-                            }
-                        };
-                        filtered
-                            .iter()
-                            .any(|ev| effect_event_method_args_eq(ev, &expected))
-                    };
-                    Ok(VmVerifyEval::Value(Value::Bool(hit)))
-                }
-                _ => Ok(helper_result),
-            }
-        }
-        _ => Ok(helper_result),
-    }
-}
-
-/// Oracle v1: recognise one `<scope>.<method>(N)` stage where method
-/// must match `expected` — used by the generic tree-nav peeler to
-/// strip `.group(N)` / `.branch(idx)` stages. Argument must be a
-/// literal non-negative Int.
-fn peel_named_int_stage<'a>(
-    expr: &'a Spanned<Expr>,
-    expected: &str,
-) -> Option<(&'a Spanned<Expr>, u32)> {
-    let Expr::FnCall(callee, args) = &expr.node else {
-        return None;
-    };
-    let Expr::Attr(base, method) = &callee.node else {
-        return None;
-    };
-    if method != expected || args.len() != 1 {
-        return None;
-    }
-    match &args[0].node {
-        Expr::Literal(aver::ast::Literal::Int(n)) if *n >= 0 => Some((base.as_ref(), *n as u32)),
-        _ => None,
-    }
-}
-
-/// Oracle v1: peel the optional `.group(N).branch(idx)` tree-nav
-/// stages sitting between `.trace` and a terminal projection method.
-/// Returns the remaining expression (pointing at `fn_call.trace`)
-/// plus the group / branch indices the user requested.
-///
-/// Supported shapes:
-///   fn_call.trace                        → (expr, None, None)
-///   fn_call.trace.group(N)               → (expr, Some(N), None)
-///   fn_call.trace.group(N).branch(idx)   → (expr, Some(N), Some(idx))
-///
-/// `.branch(idx)` without a preceding `.group(N)` is rejected by the
-/// normal projection path — branches only make sense inside a group.
-fn peel_tree_nav_stages(expr: &Spanned<Expr>) -> (&Spanned<Expr>, Option<u32>, Option<u32>) {
-    // Try `.branch(idx)` first (innermost if present).
-    if let Some((after_branch, idx)) = peel_named_int_stage(expr, "branch")
-        && let Some((after_group, gid)) = peel_named_int_stage(after_branch, "group")
-    {
-        return (after_group, Some(gid), Some(idx));
-    }
-    if let Some((after_group, gid)) = peel_named_int_stage(expr, "group") {
-        return (after_group, Some(gid), None);
-    }
-    (expr, None, None)
-}
-
-/// Oracle v1: detect a trace-projection LHS shape — `.trace`,
-/// `.trace.length()`, `.trace.event(k)`, `.trace.contains(_)`. Used to
-/// decide whether a failing assertion should be decorated with the
-/// recorded trace events (not meaningful for `.result` or non-trace
-/// assertions).
-fn is_trace_projection_shape(expr: &Spanned<Expr>) -> bool {
-    match &expr.node {
-        Expr::Attr(inner, field) if field == "trace" => {
-            matches!(&inner.node, Expr::FnCall(_, _))
-        }
-        Expr::FnCall(callee, _) => {
-            let Expr::Attr(inner_attr, method) = &callee.node else {
-                return false;
-            };
-            if !matches!(method.as_str(), "length" | "event" | "contains") {
-                return false;
-            }
-            let (scope, _, _) = peel_tree_nav_stages(inner_attr);
-            let Expr::Attr(fn_call_box, trace_field) = &scope.node else {
-                return false;
-            };
-            trace_field == "trace" && matches!(&fn_call_box.node, Expr::FnCall(_, _))
-        }
-        _ => false,
-    }
-}
-
-/// Oracle v1: render the trace-events buffer for failure diagnostics.
-/// The suffix is appended to `actual` so a failing
-/// `.trace.contains(Http.get("/x")) => true` shows exactly which
-/// events the LHS emitted, turning the default `actual: false` into
-/// something the user can act on.
-fn format_trace_events_tail(events: &[Value]) -> String {
-    if events.is_empty() {
-        return "  (trace: [] — no effects recorded)".to_string();
-    }
-    let rendered: Vec<String> = events.iter().map(aver_repr).collect();
-    format!("  (trace: [{}])", rendered.join(", "))
-}
-
-/// Oracle v1: build a Trace record wrapping the collected events.
-fn build_trace_record(collected: &[Value]) -> Value {
-    Value::Record {
-        type_name: aver::types::trace::TYPE_NAME.to_string(),
-        fields: vec![(
-            aver::types::trace::FIELD_EVENTS.to_string(),
-            aver::value::list_from_vec(collected.to_vec()),
-        )]
-        .into(),
-    }
-}
-
-/// Oracle v1: treat an effect-method call in Trace.contains-argument
-/// position as a structural event literal. `Console.print("rolled 4")`
-/// becomes `EffectEvent{method: "Console.print", args: ["rolled 4"]}`.
-/// Currently supports literal-arg calls only (string, int, float, bool,
-/// interpolated strings with literal parts) — richer arg shapes follow
-/// when event construction has its own elaboration pass.
-/// Oracle v1: equality on EffectEvent records ignoring `path`. Used by
-/// `.trace.contains(Effect.method(...))` — the user-authored literal
-/// does not carry a structural path; matching intentionally ignores
-/// it so "this event fired somewhere" is expressible.
-fn effect_event_method_args_eq(recorded: &Value, expected: &Value) -> bool {
-    let Value::Record {
-        type_name: rt,
-        fields: rf,
-    } = recorded
-    else {
-        return false;
-    };
-    let Value::Record {
-        type_name: et,
-        fields: ef,
-    } = expected
-    else {
-        return false;
-    };
-    if rt != et {
-        return false;
-    }
-    let pick = |fields: &[(String, Value)], name: &str| -> Option<Value> {
-        fields
-            .iter()
-            .find(|(n, _)| n == name)
-            .map(|(_, v)| v.clone())
-    };
-    let r_method = pick(rf, aver::types::effect_event::FIELD_METHOD);
-    let e_method = pick(ef, aver::types::effect_event::FIELD_METHOD);
-    let r_args = pick(rf, aver::types::effect_event::FIELD_ARGS);
-    let e_args = pick(ef, aver::types::effect_event::FIELD_ARGS);
-    r_method == e_method && r_args == e_args
-}
-
-/// Oracle v1: recognise a bare effect-method reference like
-/// `Console.print` (no parens) and return `"Console.print"`. Only
-/// classified effect methods count — other `Namespace.name` shapes
-/// (user types, stdlib namespaces) are rejected so the method-only
-/// `.contains(X)` branch doesn't accidentally match random refs.
-fn effect_method_ref_name(expr: &Spanned<Expr>) -> Option<String> {
-    let (ns, method) = match &expr.node {
-        Expr::Attr(ns, method) => {
-            let Expr::Ident(ns_name) = &ns.node else {
-                return None;
-            };
-            (ns_name.clone(), method.clone())
-        }
-        Expr::Resolved { name, .. } => {
-            let (ns, method) = name.rsplit_once('.')?;
-            (ns.to_string(), method.to_string())
-        }
-        _ => return None,
-    };
-    let full = format!("{}.{}", ns, method);
-    if aver::types::checker::effect_classification::is_classified(&full) {
-        Some(full)
-    } else {
-        None
-    }
-}
-
-/// Oracle v1: true when the recorded EffectEvent record has
-/// `method == expected_name`. Used by `.contains(Effect.method)`'s
-/// method-only branch.
-fn effect_event_has_method(recorded: &Value, expected_name: &str) -> bool {
-    let Value::Record { type_name, fields } = recorded else {
-        return false;
-    };
-    if type_name != aver::types::effect_event::TYPE_NAME {
-        return false;
-    }
-    fields.iter().any(|(name, val)| {
-        name == aver::types::effect_event::FIELD_METHOD
-            && matches!(val, Value::Str(s) if s == expected_name)
-    })
-}
-
-fn expr_to_effect_event(expr: &Spanned<Expr>) -> Option<Value> {
-    let Expr::FnCall(callee, args) = &expr.node else {
-        return None;
-    };
-    let Expr::Attr(ns, method) = &callee.node else {
-        return None;
-    };
-    let Expr::Ident(ns_name) = &ns.node else {
-        return None;
-    };
-    let full_name = format!("{}.{}", ns_name, method);
-    let arg_vals: Vec<Value> = args.iter().filter_map(literal_expr_to_value).collect();
-    if arg_vals.len() != args.len() {
-        return None;
-    }
-    Some(Value::Record {
-        type_name: aver::types::effect_event::TYPE_NAME.to_string(),
-        fields: vec![
-            (
-                aver::types::effect_event::FIELD_METHOD.to_string(),
-                Value::Str(full_name),
-            ),
-            (
-                aver::types::effect_event::FIELD_ARGS.to_string(),
-                aver::value::list_from_vec(arg_vals),
-            ),
-        ]
-        .into(),
-    })
-}
-
-fn literal_expr_to_value(expr: &Spanned<Expr>) -> Option<Value> {
-    match &expr.node {
-        Expr::Literal(aver::ast::Literal::Int(i)) => Some(Value::Int(*i)),
-        Expr::Literal(aver::ast::Literal::Float(f)) => Some(Value::Float(*f)),
-        Expr::Literal(aver::ast::Literal::Str(s)) => Some(Value::Str(s.clone())),
-        Expr::Literal(aver::ast::Literal::Bool(b)) => Some(Value::Bool(*b)),
-        Expr::Literal(aver::ast::Literal::Unit) => Some(Value::Unit),
-        Expr::InterpolatedStr(parts) => {
-            // Supports interpolations whose Parsed segments resolve to
-            // scalar literals (direct or via local-binding substitution
-            // at parse time). E.g. with local `x = 4`, the string
-            // `"rolled {x}"` becomes `"rolled 4"` — the Parsed segment
-            // carries `Literal(Int(4))` after substitution. Richer
-            // shapes (calls, record access, etc.) still defer to
-            // verify-context evaluation.
-            let mut s = String::new();
-            for part in parts {
-                match part {
-                    aver::ast::StrPart::Literal(lit) => s.push_str(lit),
-                    aver::ast::StrPart::Parsed(inner) => {
-                        let v = literal_expr_to_value(inner)?;
-                        s.push_str(&interp_value_to_str(&v)?);
-                    }
-                }
-            }
-            Some(Value::Str(s))
-        }
-        _ => None,
-    }
-}
-
-/// Oracle v1: render a scalar Value as it would appear in an Aver
-/// string interpolation (`"x = {v}"`). Mirrors the runtime's
-/// interp rendering for Int / Float / Bool / Str / Unit. Used only
-/// from `literal_expr_to_value` for trace event-literal strings; more
-/// complex Values aren't expected here.
-fn interp_value_to_str(value: &Value) -> Option<String> {
-    match value {
-        Value::Str(s) => Some(s.clone()),
-        Value::Int(i) => Some(i.to_string()),
-        Value::Float(f) => Some(f.to_string()),
-        Value::Bool(b) => Some(b.to_string()),
-        Value::Unit => Some(String::new()),
-        _ => None,
-    }
-}
-
-fn vm_call_verify_helper(machine: &mut vm::VM, fn_name: &str) -> Result<VmVerifyEval, String> {
-    let value = machine
-        .run_named_function(fn_name, &[])
-        .map_err(|e| e.to_string())?
-        .to_value(&machine.arena);
-
-    match value {
-        Value::Ok(inner) => Ok(VmVerifyEval::Value(*inner)),
-        Value::Err(inner) => Ok(VmVerifyEval::ErrProp(*inner)),
-        other => Err(format!(
-            "verify helper '{}' returned unexpected shape: {}",
-            fn_name,
-            aver_repr(&other)
-        )),
-    }
-}
-
-fn vm_call_guard_helper(machine: &mut vm::VM, fn_name: &str) -> Result<Value, String> {
-    machine
-        .run_named_function(fn_name, &[])
-        .map_err(|e| e.to_string())
-        .map(|value| value.to_value(&machine.arena))
-}
-
-/// Oracle v1: for a verify-law case, collect the oracle-stub mapping
-/// the VM should install before evaluating LHS/RHS.
-///
-/// For each given `name: Effect.method = [stub_expr, ...]`, the case's
-/// expansion picks one element of the list into `case_givens[idx]`. We
-/// look up the classified effect (via the law's original givens) and
-/// resolve the case-specific stub expression to a VM fn_id (must be a
-/// bare identifier). Only classified generative / snapshot / gen+output
-/// effects produce an entry; output-only / unclassified effects are
-/// skipped (they have no oracle to stub).
-/// Oracle v1: flatten a dotted `Attr(Attr(Ident, "A"), "B")` chain
-/// to the module-qualified name `"Ident.A.B"`. Returns `None` if the
-/// expression isn't a pure attr chain rooted in an Ident — anything
-/// more complex can't be a stub reference.
-fn dotted_path_from_expr(expr: &Expr) -> Option<String> {
-    match expr {
-        Expr::Ident(s) => Some(s.clone()),
-        Expr::Resolved { name, .. } => Some(name.clone()),
-        Expr::Attr(obj, field) => {
-            let head = dotted_path_from_expr(&obj.node)?;
-            Some(format!("{}.{}", head, field))
-        }
-        _ => None,
-    }
-}
-
-fn build_case_oracle_stubs(
-    machine: &vm::VM,
-    block: &VerifyBlock,
-    case_idx: usize,
-) -> std::collections::HashMap<String, u32> {
-    let mut out = std::collections::HashMap::new();
-    // Law-form stores givens inside VerifyKind::Law; cases-form (used for
-    // `verify fn trace` with top-level givens) stores them on
-    // block.cases_givens. Both shapes funnel into the same mapping.
-    let givens_source: &[aver::ast::VerifyGiven] = match &block.kind {
-        VerifyKind::Law(law) => law.givens.as_slice(),
-        VerifyKind::Cases => block.cases_givens.as_slice(),
-    };
-    if givens_source.is_empty() {
-        return out;
-    }
-    let Some(case_bindings) = block.case_givens.get(case_idx) else {
-        return out;
-    };
-    for given in givens_source {
-        let Some(classification) =
-            aver::types::checker::effect_classification::classify(&given.type_name)
-        else {
-            continue;
-        };
-        use aver::types::checker::effect_classification::EffectDimension;
-        if matches!(classification.dimension, EffectDimension::Output) {
-            continue;
-        }
-        // Find this case's binding for the given's name.
-        let Some((_, value_expr)) = case_bindings.iter().find(|(n, _)| n == &given.name) else {
-            continue;
-        };
-        // Oracle v1: the sample value must resolve to a top-level Aver
-        // fn. Accept bare Ident / Resolved (same module) AND dotted
-        // `Attr` chains like `Helpers.stubArgs` (module-qualified
-        // reference). Silently ignoring the Attr case would leave the
-        // real effect un-stubbed and the verify runner would compare
-        // against the oracle-call RHS while LHS still hit the real
-        // effect — a very subtle semantic hole.
-        let stub_name = match &value_expr.node {
-            Expr::Ident(s) => s.clone(),
-            Expr::Resolved { name, .. } => name.clone(),
-            _ => {
-                let Some(dotted) = dotted_path_from_expr(&value_expr.node) else {
-                    continue;
-                };
-                dotted
-            }
-        };
-        let Some(fn_id) = machine.find_fn_id(&stub_name) else {
-            continue;
-        };
-        out.insert(given.type_name.clone(), fn_id);
-    }
-    out
-}
-
-fn run_verify_vm(plan: &VmVerifyPlan, machine: &mut vm::VM) -> VerifyResult {
-    use aver::checker::{VerifyCaseOutcome, VerifyCaseResult, VerifyLawContext};
-
-    let block = &plan.block;
-    let mut passed = 0;
-    let mut failed = 0;
-    let mut skipped = 0;
-    let mut failures = Vec::new();
-    let mut case_results = Vec::new();
-    let is_law = matches!(block.kind, VerifyKind::Law(_));
-    let case_total = block.cases.len();
-
-    let law_context_template = if let VerifyKind::Law(law) = &block.kind {
-        Some(format!(
-            "{} == {}",
-            expr_to_str(&law.lhs),
-            expr_to_str(&law.rhs)
-        ))
-    } else {
-        None
-    };
-
-    for (idx, ((left_expr, right_expr), case_fns)) in
-        block.cases.iter().zip(&plan.cases).enumerate()
-    {
-        let case_str = format!("{} == {}", expr_to_str(left_expr), expr_to_str(right_expr));
-        let span = block.case_spans.get(idx).cloned();
-        let failure_case = if is_law {
-            format!("case {}/{} [{}]", idx + 1, case_total, case_str)
-        } else {
-            case_str.clone()
-        };
-
-        let law_context = if let VerifyKind::Law(_) = &block.kind {
-            let givens: Vec<(String, String)> = block
-                .case_givens
-                .get(idx)
-                .map(|gs| {
-                    gs.iter()
-                        .map(|(name, expr)| (name.clone(), expr_to_str(expr)))
-                        .collect()
-                })
-                .unwrap_or_default();
-            Some(VerifyLawContext {
-                givens,
-                law_expr: law_context_template.clone().unwrap_or_default(),
-            })
-        } else {
-            None
-        };
-
-        // Check law sample guard
-        if let Some(guard_name) = &case_fns.guard {
-            match vm_call_guard_helper(machine, guard_name) {
-                Ok(Value::Bool(true)) => {}
-                Ok(Value::Bool(false)) => {
-                    skipped += 1;
-                    case_results.push(VerifyCaseResult {
-                        outcome: VerifyCaseOutcome::Skipped,
-                        span,
-                        case_expr: case_str,
-                        case_index: idx,
-                        case_total,
-                        law_context,
-                    });
-                    continue;
-                }
-                Ok(Value::Err(err_val)) => {
-                    failed += 1;
-                    let err_repr = format!("Result.Err({})", aver_repr(&err_val));
-                    failures.push((failure_case, String::new(), err_repr.clone()));
-                    case_results.push(VerifyCaseResult {
-                        outcome: VerifyCaseOutcome::UnexpectedErr { err_repr },
-                        span,
-                        case_expr: case_str,
-                        case_index: idx,
-                        case_total,
-                        law_context,
-                    });
-                    continue;
-                }
-                Ok(other) => {
-                    failed += 1;
-                    let error = format!("when produced {}, expected Bool", aver_repr(&other));
-                    failures.push((failure_case, "Bool".to_string(), error.clone()));
-                    case_results.push(VerifyCaseResult {
-                        outcome: VerifyCaseOutcome::RuntimeError { error },
-                        span,
-                        case_expr: case_str,
-                        case_index: idx,
-                        case_total,
-                        law_context,
-                    });
-                    continue;
-                }
-                Err(e) => {
-                    failed += 1;
-                    let error = format!("guard error: {}", e);
-                    failures.push((failure_case, String::new(), error.clone()));
-                    case_results.push(VerifyCaseResult {
-                        outcome: VerifyCaseOutcome::RuntimeError { error },
-                        span,
-                        case_expr: case_str,
-                        case_index: idx,
-                        case_total,
-                        law_context,
-                    });
-                    continue;
-                }
-            }
-        }
-
-        // Oracle v1: for each given in the law whose type is a classified
-        // effect method, install a map from the effect name to the stub
-        // fn_id chosen for this case. The VM's oracle-stub hook will
-        // intercept classified-effect calls under LHS and re-dispatch
-        // them to the stub with a prepended (BranchPath.root, counter).
-        let oracle_stubs = build_case_oracle_stubs(machine, block, idx);
-        let has_stubs = !oracle_stubs.is_empty();
-        if has_stubs {
-            machine.install_oracle_stubs(oracle_stubs);
-        }
-
-        // Oracle v1: for trace-aware verify blocks, start per-case trace
-        // event collection. The collector lives on VmRuntime and captures
-        // every classified-effect dispatch during LHS eval. Events are
-        // taken (and cleared) before RHS eval so only LHS emissions are
-        // visible to subsequent `.trace.*` projections.
-        //
-        // Also set the helper-boundary root fn_id — only direct
-        // emissions by the fn under `verify <fn> trace` contribute.
-        // Emissions from helpers that fn calls are ghost state and
-        // don't clutter `.trace.contains(...)` / `.event(k)` /
-        // `.length()` queries.
-        if block.trace {
-            machine.start_trace_collection();
-            let root_fn_id = machine.find_fn_id(&block.fn_name);
-            machine.set_trace_root_fn_id(root_fn_id);
-        }
-
-        let left_result = vm_call_verify_helper(machine, &case_fns.left);
-        let (lhs_trace_events, lhs_trace_coords): (Vec<Value>, Vec<aver::vm::runtime::TraceCoord>) =
-            if block.trace {
-                machine.take_trace_events_with_coords()
-            } else {
-                (Vec::new(), Vec::new())
-            };
-
-        // Oracle v1: if the LHS was a trace-contents projection
-        // (`fn.trace.contains(...)` / `.event(k)` / `.length()` or a
-        // tree-nav chain `.trace.group(N).*`), the helper's returned
-        // value is a sentinel — compute the real projection value
-        // here from the collected events + coords.
-        let left_result = apply_trace_projection(
-            left_result,
-            &block.cases[idx].0,
-            &lhs_trace_events,
-            &lhs_trace_coords,
-            block.trace,
-        );
-
-        let right_result = vm_call_verify_helper(machine, &case_fns.right);
-
-        if has_stubs {
-            machine.clear_oracle_stubs();
-        }
-
-        match (left_result, right_result) {
-            (Ok(VmVerifyEval::Value(left_val)), Ok(VmVerifyEval::Value(right_val))) => {
-                if left_val == right_val {
-                    passed += 1;
-                    case_results.push(VerifyCaseResult {
-                        outcome: VerifyCaseOutcome::Pass,
-                        span,
-                        case_expr: case_str,
-                        case_index: idx,
-                        case_total,
-                        law_context,
-                    });
-                } else {
-                    failed += 1;
-                    let expected = aver_repr(&right_val);
-                    let mut actual = aver_repr(&left_val);
-                    // Oracle v1: for trace-projection LHS shapes, append
-                    // the recorded events so a failing
-                    // `.trace.contains(...) => true` (actual: false) is
-                    // debuggable at a glance — the user sees which events
-                    // actually fired.
-                    if block.trace && is_trace_projection_shape(&block.cases[idx].0) {
-                        actual.push_str(&format_trace_events_tail(&lhs_trace_events));
-                    }
-                    failures.push((failure_case, expected.clone(), actual.clone()));
-                    case_results.push(VerifyCaseResult {
-                        outcome: VerifyCaseOutcome::Mismatch { expected, actual },
-                        span,
-                        case_expr: case_str,
-                        case_index: idx,
-                        case_total,
-                        law_context,
-                    });
-                }
-            }
-            (Ok(VmVerifyEval::ErrProp(err_val)), _) | (_, Ok(VmVerifyEval::ErrProp(err_val))) => {
-                failed += 1;
-                let err_repr = format!("Result.Err({})", aver_repr(&err_val));
-                failures.push((failure_case, String::new(), err_repr.clone()));
-                case_results.push(VerifyCaseResult {
-                    outcome: VerifyCaseOutcome::UnexpectedErr { err_repr },
-                    span,
-                    case_expr: case_str,
-                    case_index: idx,
-                    case_total,
-                    law_context,
-                });
-            }
-            (Err(e), _) | (_, Err(e)) => {
-                failed += 1;
-                let error = e.to_string();
-                failures.push((failure_case, String::new(), error.clone()));
-                case_results.push(VerifyCaseResult {
-                    outcome: VerifyCaseOutcome::RuntimeError { error },
-                    span,
-                    case_expr: case_str,
-                    case_index: idx,
-                    case_total,
-                    law_context,
-                });
-            }
-        }
-    }
-
-    let block_label = match &block.kind {
-        VerifyKind::Law(law) => format!("{} spec {}", block.fn_name, law.name),
-        VerifyKind::Cases => block.fn_name.clone(),
-    };
-    VerifyResult {
-        fn_name: block.fn_name.clone(),
-        block_label,
-        passed,
-        failed,
-        skipped,
-        case_results,
-        failures,
-    }
-}
-
-fn run_verify_for_items_vm(
-    mut items: Vec<TopLevel>,
-    module_root: &str,
-    source_file: &str,
-) -> Result<Vec<VerifyResult>, String> {
-    tco::transform_program(&mut items);
-
-    let tc_result = run_type_check_full(&items, Some(module_root));
-    if !tc_result.errors.is_empty() {
-        return Err(format_type_errors(&tc_result.errors));
-    }
-
-    let verify_blocks = merge_verify_blocks(&items);
-    if verify_blocks.is_empty() {
-        return Ok(vec![]);
-    }
-
-    let plans = build_verify_vm_plans(&mut items, &verify_blocks);
-    resolver::resolve_program(&mut items);
-
-    let mut arena = Arena::new();
-    vm::register_service_types(&mut arena);
-    let (code, globals) =
-        vm::compile_program_with_modules(&items, &mut arena, Some(module_root), source_file)
-            .map_err(|e| format!("VM compile error: {}", e))?;
-    let mut machine = vm::VM::new(code, globals, arena);
-    apply_runtime_policy_to_vm(&mut machine, module_root)?;
-
-    let mut results = Vec::new();
-    for plan in &plans {
-        results.push(run_verify_vm(plan, &mut machine));
-    }
-    Ok(results)
-}
-
 struct VerifyFileResult {
     path: String,
     source: String,
@@ -3480,8 +2520,14 @@ fn run_verify_for_file(
     let units = collect_check_units(file, module_root, deps)?;
     let mut file_results = Vec::new();
 
+    let config = load_runtime_policy(module_root)?;
     for (path, source, items) in units {
-        let blocks = run_verify_for_items_vm(items, module_root, &path)?;
+        let blocks = aver::diagnostics::vm_verify::run_verify_for_items_vm(
+            items,
+            config.clone(),
+            Some(module_root),
+            &path,
+        )?;
         file_results.push(VerifyFileResult {
             path,
             source,
